Move utility methods which operate based on abstract class assumptions into the class itself.
