+++ src/genjax/_src/console.py
@@ -20,7 +20,7 @@ from rich import traceback
 from rich.console import Console
 
 from genjax._src.checkify import no_checkify, yes_checkify
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 from genjax._src.core.typing import Bool, Dict
 
 ###################
