+++ src/genjax/_src/core/datatypes/trie.py
@@ -19,7 +19,7 @@ import rich
 import genjax._src.core.pretty_printing as gpp
 from genjax._src.core.datatypes.hashable_dict import HashableDict, hashable_dict
 from genjax._src.core.pretty_printing import CustomPretty
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 
 ########
 # Trie #
