+++ src/genjax/_src/core/serialization/pickle.py
@@ -16,7 +16,7 @@
 import dill as pickle
 
 from genjax._src.core.datatypes.generative import Trace
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 from genjax._src.core.serialization.backend import SerializationBackend
 from genjax._src.core.typing import Any, dispatch
