+++ src/genjax/_src/generative_functions/combinators/vector/unfold_combinator.py
@@ -40,7 +40,7 @@ from genjax._src.core.interpreters.incremental import (
     tree_diff_primal,
     tree_diff_unknown_change,
 )
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 from genjax._src.core.typing import (
     Any,
     FloatArray,
