+++ src/genjax/_src/generative_functions/combinators/vector/vector_datatypes.py
@@ -29,9 +29,7 @@ from genjax._src.core.datatypes.generative import (
 )
 from genjax._src.core.datatypes.hashable_dict import hashable_dict
 from genjax._src.core.datatypes.trie import Trie
-from genjax._src.core.pytree.checks import (
-    static_check_tree_leaves_have_matching_leading_dim,
-)
+from genjax._src.core.pytree import Pytree
 from genjax._src.core.typing import (
     Any,
     Dict,
@@ -58,7 +56,9 @@ class IndexedSelection(Selection):
     inner: Selection
 
     def __post_init__(self):
-        static_check_tree_leaves_have_matching_leading_dim((self.inner, self.indices))
+        Pytree.static_check_tree_leaves_have_matching_leading_dim(
+            (self.inner, self.indices)
+        )
 
     @dispatch
     def has_addr(self, addr: IntArray):
@@ -201,7 +201,7 @@ class VectorChoiceMap(ChoiceMap):
     inner: Any
 
     def __post_int__(self):
-        static_check_tree_leaves_have_matching_leading_dim(self.inner)
+        Pytree.static_check_tree_leaves_have_matching_leading_dim(self.inner)
 
     def is_empty(self):
         return self.inner.is_empty()
@@ -212,7 +212,7 @@ class VectorChoiceMap(ChoiceMap):
         selection: IndexedSelection,
     ) -> ChoiceMap:
         inner = self.inner.filter(selection.inner)
-        dim = static_check_tree_leaves_have_matching_leading_dim(inner)
+        dim = Pytree.static_check_tree_leaves_have_matching_leading_dim(inner)
         check = selection.indices <= dim
         idxs = check * selection.indices
         return IndexedChoiceMap(
@@ -229,14 +229,14 @@ class VectorChoiceMap(ChoiceMap):
     def get_selection(self):
         subselection = self.inner.get_selection()
         # Static: get the leading dimension size value.
-        dim = static_check_tree_leaves_have_matching_leading_dim(
+        dim = Pytree.static_check_tree_leaves_have_matching_leading_dim(
             self.inner,
         )
         return IndexedSelection(jnp.arange(dim), subselection)
 
     @dispatch
     def has_submap(self, addr: IntArray):
-        dim = static_check_tree_leaves_have_matching_leading_dim(
+        dim = Pytree.static_check_tree_leaves_have_matching_leading_dim(
             self.inner,
         )
         return addr < dim
@@ -244,7 +244,7 @@ class VectorChoiceMap(ChoiceMap):
     @dispatch
     def has_submap(self, addr: Tuple):
         (idx, *addr) = addr
-        dim = static_check_tree_leaves_have_matching_leading_dim(
+        dim = Pytree.static_check_tree_leaves_have_matching_leading_dim(
             self.inner,
         )
         return jnp.logical_and(idx < dim, self.inner.has_submap(tuple(addr)))
@@ -256,7 +256,7 @@ class VectorChoiceMap(ChoiceMap):
 
     @dispatch
     def get_submap(self, idx: Int):
-        dim = static_check_tree_leaves_have_matching_leading_dim(
+        dim = Pytree.static_check_tree_leaves_have_matching_leading_dim(
             self.inner,
         )
         check = idx < dim
@@ -266,7 +266,7 @@ class VectorChoiceMap(ChoiceMap):
 
     @dispatch
     def get_submap(self, idx: IntArray):
-        dim = static_check_tree_leaves_have_matching_leading_dim(
+        dim = Pytree.static_check_tree_leaves_have_matching_leading_dim(
             self.inner,
         )
         check = idx < dim
