+++ src/genjax/_src/generative_functions/distributions/custom/discrete_hmm.py
@@ -19,7 +19,7 @@ import numpy as np
 from scipy.linalg import circulant
 from tensorflow_probability.substrates import jax as tfp
 
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 from genjax._src.core.typing import FloatArray, IntArray, PRNGKey
 from genjax._src.generative_functions.distributions.distribution import Distribution
