+++ src/genjax/_src/generative_functions/static/static_gen_fn.py
@@ -23,7 +23,7 @@ from genjax._src.core.datatypes.generative import (
 )
 from genjax._src.core.interpreters.incremental import static_check_tree_leaves_diff
 from genjax._src.core.interpreters.staging import stage
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 from genjax._src.core.typing import (
     Any,
     Callable,
