+++ src/genjax/_src/inference/core.py
@@ -21,7 +21,7 @@ from genjax._src.core.datatypes.generative import (
     GenerativeFunction,
     Selection,
 )
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 from genjax._src.core.typing import (
     Any,
     Callable,
