+++ src/genjax/_src/information/sdos.py
@@ -24,7 +24,7 @@ from genjax._src.core.datatypes.generative import (
     GenerativeFunction,
     Selection,
 )
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 from genjax._src.core.typing import Int, PRNGKey, Tuple
 from genjax._src.inference.core import Marginal, Target
