+++ src/genjax/core/pytree.py
@@ -12,11 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from genjax._src.core.pytree.closure import DynamicClosure, dynamic_closure
-from genjax._src.core.pytree.pytree import Pytree
+from genjax._src.core.pytree import Pytree
 
 __all__ = [
-    "DynamicClosure",
-    "dynamic_closure",
     "Pytree",
 ]
